Fix date handling
We have a problem when converting a date into the current
system timezone: we have been grabbing the timezone based on the
system at the current date, but we actually want the system timezone
at the original date.

Suppose our machine's timezone is set to Pacific/Auckland and we run
the following:

    --date 2022-04-10 -s america/new_york 08:30

The exact output of this will depend on if we are currently observing
daylight savings time (NZDT vs. NZST), but it really shouldn't! We
should be outputing what Pacific/Auckland would have given at that
specific time. This is in line with the rest of the application
i.e. we are interested in time associated to physical locations
(tz_database labels), not arbitrary labels (NZDT vs. NZST).

To support this, we make several significant changes:

- When converting time, we no longer use the current system timezone,
  as it may be wrong. We use getTimeZone, which takes in a UTCTime
  param to derive the TimeZone we need.

- When reading a time, we take care to read it in the timezone at the
  given --date, if we receive one.

- Time reading is much more explicit, for ease of understanding.

- The old style of concatenating string together then parsing is mostly
  replaced with parsing simpler types (e.g. TimeOfDay, Day), then
  combining as necessary. Again, this is easier to understand.

- If --date is not given, we use the current date, as determined by the
  source. The previous logic -- arbitrarily choosing the unix epoch --
  was not actually useful, and this way is easier to understand:
  we are either given a date, which we use, or we use the current date.
  This also means we no longer need the 'today' option for --date,
  so it has been removed.